requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart
psycopg2-binary
python-dotenv
httpx
pillow
pydantic>=2.0
cloudinary
numpy
authlib
itsdangerous
python-jose[cryptography]
bcrypt
transformers
torch

